Add makefile
The purpose of the makefile is to allow users to compile and create an executable file for their specific machine all in one process by invoking the command `make` from their terminal/command line. Previously, it was assumed the executable that was included with the initial version of TicTacToe would run on any device, but after some research I found out that was not the case.