Rollout of v1.3.0
What's new in Version 1.3.0: the main features are 1) the ability to play multiple games in a row without having to quit, and 2) score-keeping functionality that tracks the number of wins or ties for each player. Note: hurdles were cleared by getting the `resetBoard()` function to properly reassign number values to each position of the board after a game ends. This was the biggest problem faced since the development of the game and I am happy to have solved the problem. Also, comments were added throughout the code for better readability.